Deploy_Apps router
- During the initial staging build we encontered a problem with having
  access to the application binary files from the seed environment. For
example integration will be the seed environment of staging. Due to this
conundrum, we created a temporary solution to allow the router
application (key to all the other applications) to run as a container
using DockerHub. This is no longer required. However, this is an option
that has been proven to work. Hence, we have mearly changed the option
from true to false.

Solo: @suthagarht