Add whitehall 5xx analysis grafana dashboard
These exist on production and staging but were manually provisioned so
they were a) different between staging and prod, b) broken because they
used the old style prefixes for querying kibana.  Moving the definition
into puppet should avoid those discrepancies in the future.

We add it as a flat file rather than an ERB template because there's
currently no need to use any ERB templating features to define this
dashboard.