Add redis-tools package to monitoring machines
In Carrenza we can ssh onto the redis machines and run redis-cli to
run queries directly against redis.  On AWS we can't ssh onto the
elasticache redis instances so we can't do this.  We don't have redis-cli
on many boxes, so adding it to the monitoring ones seems like a good
first step.