Increase memory usage threshold for gov frontend
We can see that the average memory usage for government-frontend hovers around
660 MB - 740 MB during a day:

https://graphite.publishing.service.gov.uk/render/?width=586&height=308&_salt=1520870375.531&target=movingAverage(frontend-?_frontend.processes-app-government-frontend.ps_data%2C1000)

The current warning threshold is set very close to that (warning: 900MB, critical: 1000MB). Once it reaches the critical threshold Icinga will restart all the apps processees.

We're therefore increasing the threshold to be double the current one, both for warning level and for critical.

The machines have 8GB of RAM each, so we are well below the full capacity of the boxes:

https://github.com/alphagov/govuk-provisioning/blob/master/vcloud-launcher/production_carrenza/frontend.yaml#L106