Persist python dependencies between jobs
I originally set this up to create virtualenvs in the workspace,
to isolate the job from the rest of the system as much as possible.
However, on this environment, it takes 15 minutes to install the python
dependencies, and then a few seconds to run the actual code I want to
run.

By default, pipenv will create a virtualenv in the user's home
directory, using a name based on the Pipenv's path on the filesystem.

This means that if the job is moved or deleted, the virtualenv will be
left behind, and packages will hang around between builds. The workspace
itself will still be cleaned each time.