Incrasee keepLastValue period for rummager index size check
The data is generated by Jenkins jobs, however due to load, data points can be
delayed past the current 11  minute interval. Data should be generated every 10,
as such we are increasing the interval to 21 minutes.

As this only affects missing data points we don't feel that we are incraseing
teh risk of missin systemn issues.