Add Jenkins job to run ETL::Master process for CPM
Content Performance Manager runs the master process for ETL processing
every night on a daily basis.

We have decided to implement it in Jenkisn as it is a very simple
approach that fulfills what the product needs as of today. We might
change it into a more 12 factor approach in the long run, but it has
proved a bit tricky as we would need to handle the single exectution of
a process within a cluster.