Set up EMAIL_ALERT_AUTH_TOKEN env vars
EMAIL_ALERT_AUTH_TOKEN is a secret that is a secret that is shared
between apps to encrypt/decrpyt JWT tokens used for authenticating
users.

Email Alert API is responsible for generating and emailing tokens to
users, whereas Email Alert Frontend is responsible for decrypting it and
making use of the contents.