Fix Notify email members whitelist in Heiradata
There were a number of email-alert-api app healthcheck not ok errors
(technical failures). Specifically the email whitelist in Notify was
different to the common.yaml files in hieradata and hieradata_aws
folders.

This fix removed individuals from the whitelists in the common.yaml
files that are no longer present in the Notify whitelist.