Add Strict-Transport-Security headers in integration
This commit includes the `add-sts.conf` file in nginx config for integration, which will set the `Strict-Transport-Security` header. AWS ELBs should pass this through to the requester.

This is done separately in integration since we disable SSL completely, and a side-effect of this is to stop setting the `Strict-Transport-Security` header. SSL is disabled because SSL in integration is terminated at the AWS ELBs.