Give varnish half a gig less RAM to work with
The cache boxes are regularly hitting less than 5% free RAM in
production and staging. This is due to the slowly increasing size of
the routing table, meaning router needs more RAM. Varnish uses the
entirety of its available RAM, however it very rarely evicts objects
due to running low on memory. Almost all objects are evicted due to
expiration. This indicates that we can afford to give varnish less RAM
and not have our hit ratio suffer.

Rather than messing with the percentage of the VM’s RAM we make
available, we just take away another half gig, which is approximately
the current normal usage of a `router` process.