Enable slack notifications for Jenkins based off hieradata
Previously these were enabled by an invalid environment variable
("disabled for environment") however this caused problems in the latest
version of Jenkins slack plugin and meant it would fail to run the job.

This also removes the configuration for slack jobs which are set to send
to non-existant rooms and would therefore not have been sent.