Use an encrypted Github client secret for Jenkins
In newer versions of Jenkins the client_secret value is encrypted in the
config file. If a plain text one is added this will cause Jenkins to
re-save the file with the value encrypted which causes a Jenkins reload
configuration everytime the file is run.

By instead passing in the encrypted value we can avoid this constant
reload of configuration.