Install wal-e using python 3
The old version of WAL-E we were using depends on a broken version of the `azure` package. The newer versions depend on Python 3. Our version of puppet doesn't supply a `pip3` package provider, so I've backported it from puppet master.

The only change relevant to us is that all storage providers are now optional, hence the package being installed being changed to `wal-e[aws]`.