Increase oplog_size for api-mongos
Secondary mongos are regularly getting stuck in recovery
state in staging. On investigation the oplog size is very
marginal compared to the number of items to replicate multiplied
by the average object size.
See https://stackoverflow.com/questions/30250872/endless-recovering-state-of-secondary for
information on oplog tuning.
We've already applied a similar config for other mongo clusters.
https://github.com/alphagov/govuk-puppet/blob/2b80fc6a11cfa1c1685fcb415802b52c0eb690fd/hieradata/class/mongo.yaml#L36