Add a mount for MongoDB sync
This adds a specific mount for the MongoDB sync that occurs nightly to
copy data from production to staging/integration. We have been hitting a
problem with this recently as we have 27GB available on the os-root disk
and currently the data copy uses 26GB - so if we have a situation where
both staging and integration run at the same time the machine will run
out of disk space.

This adds a 64GB disk that is used for the purpose of copying data and
is (at least at this point in time) sufficient in size for both staging
and integration data dumps.

Previously the sync was writing to /home/deploy as the deploy user.
Instead this will write to /var/lib/mongo-sync which is owned by the
deploy user so that it can be written to.