Move mongo sync directory to mongodb::server class
In https://github.com/alphagov/govuk-puppet/pull/7573 I added an extra
mount for the sync process which used the /var/lib/mongo-sync directory.
This was only added to the api_mongo node and meant that the sync was
failing on other mongo nodes as they lacked this directory.

I have now moved it to be defined as part of mongodb::server so all
nodes which include this class will have it and the same directory can
be used throughout the sync.