Alert when at 60% disk space on mongo boxes
We learnt in the incident on the 18th May that disk space can shoot up
rapidly on mongo boxes and to avoid risks of further incidents the
recommendation was that our mongo boxes have ~50% disk space available
as leeway.

This sets alerts on all mongo data directories when they hit 60% disk
usage and goes to critical at 85%. There is the option that we think
about what each machine stores and set alerts individually for them but
I felt that added a bunch of complexity and that all our mongo boxes
could suffer the same fate if data rapidly changed.

A side effect of this is that it will produce some alerts straight away
as we have 74% disk usage in mongo-* servers and are close with 59% on
performance-mongo-*.