Fix for Sinatra app (a.k.a rummager)
Our Rummager app runs on sinatra.rb, which doesn't have a `--sandbox`
flag. We have decided to alert the user about trying to use the flag
instead of silently execute the console. The reasoning behind this is to
avoid people thinking they are in a safe session and all their changes
will be rollbacked - which will not.