Add headless Chrome and dependencies
This commit adds support for testing with headless Chrome as an alternative to PhantomJS, which has now been abandoned.

It installs the following three dependencies:

* The latest stable version of Chrome
* The Selenium standalone server, which provides programmatic access to drive Chrome
* ChromeDriver, which Selenium uses to connect to Chrome

These are installed on development VMs as well as CI machines.

Ruby projects can use the `selenium-webdriver` gem, as well as `capybara`, to run integration tests using headless Chrome.