Reorder AWS CLI arguments
Move the `--sse` argument to `aws s3 cp` to the end of the command
(after the positional args).

Reason: the `--sse` option in newer versions of the AWS CLI accept an
optional encryption type. This means it must be given at the end of the
argument list, as there would be an ambiguity between the encryption
type and the positional argument that followed (the source file for the
copy, in this case).