Don't log rotate transistion-stats cache
This stops us running logrotate on files that pre-transition-stats and
transition-stats are using on data cache files. We don't need to keep
this as logs as they are just cache for a process that is running which
deletes it's files after use. By logrotating them we have built up 90GB
of data that we can remove.

The logs themselves are already managed by logrotate.