Mark email alert checks as unstable for exit(1)
Both of these checks return an exit code 2 for when they have
definitively failed so we expect that an exit code of 1 means that an
exception has occurred which may just be a network timeout.

We'll have a freshness check in icinga to alert if these jobs don't
report a success or fail for a long time.