Start making Vagrant usable with AWS environments
The existing Vagrant configuration assumes the legacy setup of vCloud in
Carrenza. This means that quite a few code paths that run in the AWS
environments just won't work, as AWS-specific facts are not set and
Hiera data is coming from the legacy `hieradata` instead of
`hieradata_aws`.

There's more work to be done here as this code is still depending on the
`govuk-provisioning` repo being present in the parent directory.