Work around netcat-openbsd bug
`nc` commands used to send datapoints into statsd can end up pegging the
CPU and failing to exit. I observed this in Puppet master builds in
Vagrant. It wasn't clear why this doesn't happen in all cases, but it is
a known bug: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=752931

The workaround is to use `-q 1` instead of `-w 1`.

The bug is fixed upstream and the netcat in Ubuntu Bionic shouldn't be
affected (Xenial's netcat is the same version as Trusty's though).