Rewrite the doc for adding a new app
The original doc was quite verbose and missing a few changes. It seems
better to give a few rich example files, together with a list of things
to check, instead of having a section for every individual change to
make. The doc is now somewhat more Rails/PG oriented as a consequence.