Add JWS_AUTH_SECRET to collections-publisher
We want to allow people with an access_limiting token to be able to view step-by-step navigation
content published by collections-publisher on the draft stack without the need of signon account.

The first step is to create JWS_AUTH_SECRET environment variable that will be used to generate
the token.