Try removing cyclic dependency for db_admin
The puppet logs are showing a cyclic dependency involving
content_audit_tool and content_publisher when attempting to modify the
pg_hba config for the latter. It's unclear why there's a dependency
chain for these databases, but since content_publisher is new it's
possible the previous dependencies were satisfied 'in bulk' at some
point, while new dependencies should be at the bottom of the list.