Fix incorrect user for content-publisher DB
Using a combination of literal strings and variables with different
conventions for '-' and '_' made this mistake easy to make.

Instead of simply correcting the error, we've also moved similar uses of
content-publisher/content_publisher into the same class, and removed
uses of literal strings, as well as recommending others do the same.