Fix content-publisher integration DB user
The DB user should be snake case, but was previously being set using the
app_name variable, which uses hyphens. This mistake was easy to make due
to the use of literal strings and variables, which has now been fixed by
centralising all the variables and explicitly referring to them in db.