Alert on the procfile worker thread count
There have been a couple of times when the Sidekiq workers for the
Publishing API used far too many threads (one time maxing out at
1000). There should be one thread per worker, so with the current
configuration it should just be 16.

Enable this check by default for all procfile workers, as it's useful
to have coverage. Go for a high default value, as there are some
services (the Email Alert API for example) which currently use 40
threads.