Handle draft-whitehall-frontend through the whitehall-frontend
Rather than the whitehall-admin load balancer and service. Currently,
some of the resources on the draft-whitehall-frontend rendered pages
don't load, because they are loaded from the Whitehall Admin, and the
browsers Same Origin Policy doesn't allow loading these.

Serving the requests from the whitehall-frontend uses different
configuration for the assets, which use relative URLs which work.