Configure memcached to listen publicly
This will enable using memcached in a distributed way. This makes more
memory available, but the main advantage of this is that it allows
explicit cache invalidation to be done.

At the moment, you could remove entries from the cache when they
become invalid, but this would only take effect on the current
machine. When using memcached in a distributed way, the cache
invalidation would affect all machines.