Refactor Sentry config for PuppetServer
Having `$puppet_sentry_dsn` be a parameter of
`puppet::puppetserver::package` doesn't really make sense. Make it a
parameter of `puppet::puppetserver` instead.

Separate the Sentry resources into their own class, and include that if
a Sentry DSN is specified in the class parameters.

Update the rspec tests and add some new ones.

Replace the `inline_template()` that was used to write the
environment variable line with a `sprintf()` and ensure that the DSN is
double-quoted in the `/etc/default/puppetserver` file.