Rework build script for mirror repos check
Initially this job acted as a placeholder for the Icinga
service and the conditional string comparison steps didn't appear
correctly in Jenkins.
The script has been simplified so that the build is parameterised
with the STATUS param (the status of the upstream Mirror_Repositories
build) and this is used to pass or fail the build.