Jenkins Job - Mongo data-sync [WIP]
- We have a requirement to create a jenkins job that could used to
trigger the `data-sync`.

- This change addresses the following.

a) `govuk-backup` is the group that has the ownership of the backup
location. Hence, we are adding the `deploy` user to this group to
inherit the permissions.

b) The backup dumps were stored with complete permissions to the owner.
Here we are issuing write permissions to the group.

https://trello.com/c/8wzAOsrV/1314-adapt-jenkins-jobs-to-run-mongo-backup-jobs

Solo: @suthagarht