Update Node repo to access v8
This will allow us to update to version 8 of Node (8.11.3 to be exact)
which is a long-term support version of Node. Ideally we'd update to
version 10 which is the current one but this is not supported on Trusty:
https://github.com/nodesource/distributions#debian-and-ubuntu-based-distributions

This will mean that the currently installed version isn't available once
this has been added so we'll have to deploy a change to update node
shortly after.