Install yarn on dev, ci and app servers
This will allow the use of Yarn (https://yarnpkg.com/) to install Node
JS packages. This is the preferred Node package manager of Rails since
the 5.1 release of Rails https://guides.rubyonrails.org/5_1_release_notes.html#yarn-support

I expect this will only be needed on CI agents (for builds), app servers
(for building assets on a deploy) and development (for local usage) -
this contrasts with nodejs that seems to be installed on every box which
seems unnecessary. I might well be missing an important place for this
though.