Enable kernel ip_forward setting
This setting is enabled to allow Docker containers access to the outside
world [1]. The reason for adding this in is because we have found that
suddenly our docker containers are not able to build.

Investigating into why our docker containers started suddenly having
problems reveals that when Docker starts it changes this value and then
if this gets reset elsewhere the docker service will need a restart to
start working again. Changing this here stops the risk of this
occurring.

This could be changed to only run as part of govuk_docker however this
would be a messier more complicated solution.

[1]: https://github.com/moby/moby/issues/490