Allow smartanswers to run in draft preview
Trello: https://trello.com/c/W86Z7ABG/214-get-smart-answers-working-on-draft-stack

At the moment a 500 error is returned when you try to navigate the landing page of a smartanswer in draft preview.
I think this is because the smartanswers app is not running in draft.

These changes are based on the set-up for draft-government-frontend
Rather than starting up a new calculators machine, draft-smartanswers will be run on the frontend machine.