Create User - govuk-backup
- When a Dev VM is created for the first time, govuk-backup doesn't
get created by default.

- This was happening because the `govuk::deploy::setup` class doesn't
call the class that creates the `govuk-backup` user. [backup::client]

- To address this issue, we are calling the `backup::client` class
within `govuk::deploy::setup`. We are doing this at the highest level ,
because this does apply to the new AWS environments and the old Carrenza
environments.

- The original `backup::client` class doesn't explicitly create the
`govuk-backup` group. We have changed this behaviour. This was done
because
https://github.com/alphagov/govuk-puppet/blob/master/modules/govuk/manifests/deploy/setup.pp#L45
does an ordering where the `assets` and `govuk-backup` groups should be
created before we create the `govuk-backup`.

https://trello.com/c/8wzAOsrV/1314-adapt-jenkins-jobs-to-run-mongo-backup-jobs

Solo: @suthagar