Copy attachments to staging every day
I noticed that we were seeing alerts on 2nd line because this runs at
10pm Monday to Friday, which means on Monday morning it hasn't run since
Friday night and is, therefore, stale. I was going to change this to run
Sunday to Thursday but I noticed that the copy-data-to-staging runs
every day so it makes sense that this should also run at an identical
frequency.