Install foreman for ruby on Dev VM
This installs foreman with each version of Ruby in the dev VM
environment. Quite a few of our Rails apps use foreman and nearly all of
them make use of a procfile, however frequently we use the approach that is
distinctly discouraged by the foreman author by installing it in the
Gemfile.

On most systems or local machines it's really not too much of a
bother installing foreman it's something we can put in a ./startup.sh
script, however on the dev VM when gems need to be installed as root
this is more difficult - and it feels an anti-pattern to have a `sudo
gem install foreman` in a ./startup.sh script.

Now onto the actual implementation, in an ideal world this would be
handled by updating https://github.com/gds-operations/puppet-rbenv to
support a hash of gems and versions to install, however since the
version of puppet we currently use has very little loop support this
would be very difficult. Instead I've implemented this in the
govuk_rbenv as I was able to use a define as a means of processing a
loop.

This solution could hit problems if we decide we need to install
different versions of gems for different versions of ruby. In that case
it might be that we decide we want to extend
https://github.com/gds-operations/puppet-rbenv and look at defines there
that can be used for installing a particular gem.