Increase the Rummager index size monitoring time
We sometimes see this alert because the most recent index sizes
haven't yet been added to Graphite. Since we increased the number
of executors available in Jenkins (which is what populates the
index sizes) we haven't seen this be as much of an issue. However
it's not completely resolved so we can increase the time to take
graphite results from.