Alert if applications use more than 100 threads
This adds thread count monitoring for apps, similar to that already
present for procfile workers. 100 should work as a default for all
applications (asset-manager uses 98, which looks to be the highest).

Like for Sidekiq, it appears to be possible for the number of threads
to grow, and eventually cause problems. This happened recently for the
Content Performance Manager.