Monitor Dimensions and Facts via Grafana in Content Data API
This is a continuation of work in[1].

On a daily basis, we publish a few metrics via StatsD about the status
of the dimensions and facts table. This will enable the team to have a
better understanding of how the metrics and items evolve throuth time
and also detect issues / bugs.

[1]: https://trello.com/c/85P5zbF2/623-3-etl-monitor-add-data-about-dimensions