Add import job for content publisher
This commit adds a Jenkins job to stream documents from Whitehall into
Content Publisher over an SSH pipe, on integration only. This means we
can regularly test importing data as part of the lifecycle of the
application, and also preview the app with realistic data.

For some reason the combination of commands on the Whitehall machine
makes all the output go to STDERR. That and the unwanted logging sniffer
output means we use grep to filter for the lines which contain a valid
document export. Although somewhat hacky, this is only a temporary job.