Fix and standardise trigger for Data_Sync_Complete
Previously there was a trigger job for Carenza and another for AWS, both
doing the same thing. The Carenza one is more flexible as it makes use
of parameters, allowing it's re-use for both staging and integration.
I've also renamed 'api_key' to 'jenkins_api_user_password' for clarity.
Testing manually also showed the auth tokens to be unnecessary.