Export GA_UNIVERSAL_ID to publisher app
In order for some service signin pages to be cross-domain
trackable we need to append a tracking id to all form interactions.
The url in the options for the static sign in pages lives in
some YAML in publisher so at the point of (re)publishing this
content to the publishing API we need to insert the tracking ID
as a clientId=UA-xxxx-xx url parameter.