Update deployment dashboard dependent apps
These additions show 500 errors for associated applications on the relevant
deployment dashboards.  It's useful to be able to see if your deployment of
(for example) Imminence has coincided with a spike of 500 errors on Frontend.

I've added government-frontend to the rummager section in part because rummager
powers the related content section of the sidebar navigation component which is
shown.  We also ran the content pages navigation test in government frontend,
which used rummager heavily.  We've stopped this test now, but I can easily see
another incarnation being run at some point.