Pin docker version to 17.09.0-ce
This is the current version which is currently 'controlled' by
whatever is the most recent version on the apt mirror.

We want to upgrade docker to the latest 17.x version, but
we don't want to go all the way to 18.x yet, which is what
would happen if we updated the mirror.

This pins docker to the current version so we can update
the mirror and update docker in a controlled way.