Add Jenkins job for govuk-alert-tracker
We have a new app govuk-alert-tracker which runs a script that scrapes Icinga for alerts and stores it in a Google sheet. Our Platform Health Dashboard then extracts the data from the sheet to populate the dashboard metrics.

This commit adds a Jenkins job that run the rake task on the first day of each month at 7am.