Add checks for staging and integration TLS certificates
This commit adds two Icinga checks to ensure the certificates served by Fastly for our staging and integration CDN environments are valid. These certificates are the wildcard publishing certificates but they are hosted by Fastly, so whenever they are renewed, we also need to provide a copy to Fastly.

It also adds links for all of the certificate checks to a page on the developer docs with more information.