Move credentials to different directory
When running the rake task, Jenkins starts by wiping out the workspace it's
running it from. This change ensures the task runs from a subfolder, and that
credentials have been copied to the parent folder.

For context - the credentials file is created during a puppet run, which
happens every 30 mn - this will need to have happened before running the rake
task.