GovUK-Migr-Trello1618: Limit Docker Hub Usage
The docker hub credentials are not deployed to the jenkins deployer
instances in staging_aws & production_aws.

Anything to do with docker hub as part of the Deploy_APP Jenkins Job
should only be done in the integration environment.

Here we are only specifying the docket hub credential bindings for the
integration environment, so the Deploy_APP job in all other envs will
notreference these credential and thus not throw an error if the
credentials do not exist in the jenkins credentials store.

Solo: @ronocg