Add HTTP/2 preconnect header for assets
This commit adds a `Link` HTTP header pointing at the relevant assets subdomain for the environment.

This header is used by browsers that support HTTP/2 to indicate that they should immediately initiate a TCP connection to this subdomain in anticipation of downloading assets from it.

This allows the connection initiation overheads to happen while the HTML is being downloaded, so that referenced assets can then be downloaded immediately rather than waiting for the connection to be initiated.