GovUk-Migr-Trello-1626: Fix perms icinga
There is no need to lay down the nagios.cmd file using puppet.
It is a pipe that all icinga commands get sent to and when icinga starts
it will lay it down.

In order to send commands from the icinga webUI, example: command suck
as acknowledgements.

Then the nagios.cmd file will need to be executable by the www-data
group. We can make sure the nagios.cmd file ha group www-data by seting
the setgid bit on the upper directory rw.

Tested this in production_aws successfully.

Solo: @ronocg