Adding decription placeholder to Carrenza Jenkins
We have moved Jenkins to a new enviroment. This PR is to enable Jenkins to
inform users that the jobs are being actioned by a diffirent jenkins.

Fix auto load linting errors. Fix comments and move the jobs to run on AWS staging. This way we can verify the funcationality.
Resolve lint error and spacing issues cuase.