Re-order the CKAN plugins
The S3 resource uploader plugin must activate
after the broader govuk plugin, as it messes with
the resource metadata in ways which break our
XLS to CSV plugin.

The XLS to CSV plugin does its best to simulate
real manual file uploads, so the S3 plugin works
fine with them.