BAU: disable nscd for bouncer
Last week it was decided that we would enable name caching on bouncer in
aws production and aws staging. It was envisaged that this would reduce
the dns resolution requests during gor traffic replay.

But it turns out that bouncer is not doing dns lookups on the url names,
it simpy looks them up in the transition db and then sens back a 301
redirect without carrying out a dns lookup.

Since this is the case we are going to remove nscd on bouncer in aws
staging and aws production now.

Solo: @ronocg @FredericFran-GDS