Disable govuk CDN logs alert on staging
cf9fd75 moved CDN logs from the `logs-cdn` machine class to
`monitoring`, and also changed the frequency of log rotation from
daily to hourly.

On staging, traffic on the govuk CDN is so slow that frequently an
hour passes with no traffic.  In consequence, the logs are frequently
rotated between checks.  This causes a critical alert as the expected
log file does not exist.

Given that the alert thresholds for staging previouly allowed up to
two days to elapse with no logs being received, and now we only keep
logs for a couple of hours, just remove the alert.