Use copytruncate to rotate CDN logs
cf9fd75 moved CDN logs from the `logs-cdn` machine class to
`monitoring`, and also changed the frequency of log rotation from
daily to hourly.

On staging, traffic on the govuk CDN is so slow that frequently an
hour passes with no traffic.  In consequence, the logs are frequently
rotated between checks.  This causes a critical alert as the expected
log file does not exist.

The `copytruncate` option in logrotate causes it to leave the original
file in place, but truncate it.  This solves the alert problem.  As
it's the original file, we don't need to reload rsyslog, as its
current file handles will still be valid.