Upgrade nginx
This commit replaces the default Ubuntu repos with a mirrored version of the nginx master repo for nginx packages.

It also remodels the `nginx::package` class to reflect the different package layout.

`nginx-module-perl` is installed since this is the only module includes in `nginx-extras` that we actually use, for tracking request IDs across the stack.