Remove custom packages and versions
This commit removes all nginx packages and versions that override the defaults in `nginx::package`. It sets a single version using `common.yaml` and only overrides it for old Precise machines.

The original reason for these overrides was to include the nginx perl module on machines where it is used via the `nginx-extras` package. This module is now installed by default on all machines.

It also defines the apt mirror to download the packages from.