Remove use of more_set_headers
`more_set_headers` is part of a third-party nginx module which isn’t included in the default nginx builds. The newer versions of nginx we’re installing now allow `add_header` to be used for 404 responses when used with the `always` modifier, so this custom module is no longer needed.