Only keep one hour of old CDN logs
CDN logs are archived to S3.  The only reason we send them to our
monitoring server is so that we have real-time logs if necessary, as
there's a 15-minute lag on them being sent to S3 by Fastly.

Given that we're not archiving logs here, there's no more to keep more
than the bare minimum.  The previous hour's logs could be useful if
you're investigating something in real-time, so only keep one
rotation.