Adding the app_domain config for local-links-manager to use AWS
During the migration we need to test local-links-manager in AWS by having it
talk to the AWS signon service. This uses a different app_domain to Carrenza so
not only needs a new value but a conditional statement in order to determine
what VAR and VALUE is used. This commit adds the puppet code but not the new
value.