Change the website_root to remove -origin
Following on from the changes to setup Fastly with a working TLS
certificate for the www. domain for the Staging and Integration
environments, this can now be used.

Therefore, change the common hieradata for both Carrenza and AWS
environments to not contain -origin. This should propagate out to
applications through Plek.

This shouldn't affect production, as it specifies the website_root
separately.

The motivation for this is fixing the cache-clearing-service clearing
the Fastly cache in Integration, as currently it fails, as it makes
the request to www-origin, rather than the www domain.